refactor: AcceptanceTest 패키지 이동
