+++ src/test/java/kitchenpos/acceptance/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos;
+package kitchenpos.acceptance;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
