+++ src/test/java/kitchenpos/acceptance/ProductAcceptanceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos;
+package kitchenpos.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
